Getting recipes and adding to favourite

+++ app/src/main/java/com/whatthehealth/models/FoundRecipe.java
@@ -6,15 +6,15 @@ import java.util.ArrayList;
 
 public class FoundRecipe {
     @SerializedName("id")
-    private float id;
+    private Integer id;
     @SerializedName("image")
     private String image;
     @SerializedName("imageType")
     private String imageType;
     @SerializedName("likes")
-    private float likes;
+    private Integer likes;
     @SerializedName("missedIngredientCount")
-    private float missedIngredientCount;
+    private Integer missedIngredientCount;
     @SerializedName("missedIngredients")
     ArrayList < Object > missedIngredients = new ArrayList < Object > ();
     @SerializedName("title")
@@ -22,65 +22,51 @@ public class FoundRecipe {
     @SerializedName("unusedIngredients")
     ArrayList < Object > unusedIngredients = new ArrayList < Object > ();
     @SerializedName("usedIngredientCount")
-    private float usedIngredientCount;
+    private Integer usedIngredientCount;
     @SerializedName("usedIngredients")
     ArrayList < Object > usedIngredients = new ArrayList< Object >();
 
-        public float getId() {
+        public Integer getId() {
             return id;
         }
-
         public String getImage() {
             return image;
         }
-
         public String getImageType() {
             return imageType;
         }
-
-        public float getLikes() {
+        public Integer getLikes() {
             return likes;
         }
-
-        public float getMissedIngredientCount() {
+        public Integer getMissedIngredientCount() {
             return missedIngredientCount;
         }
-
         public String getTitle() {
             return title;
         }
-
         public float getUsedIngredientCount() {
             return usedIngredientCount;
         }
 
-        // Setter Methods
-
-        public void setId(float id) {
+        public void setId(Integer id) {
             this.id = id;
         }
-
         public void setImage(String image) {
             this.image = image;
         }
-
         public void setImageType(String imageType) {
             this.imageType = imageType;
         }
-
-        public void setLikes(float likes) {
+        public void setLikes(Integer likes) {
             this.likes = likes;
         }
-
-        public void setMissedIngredientCount(float missedIngredientCount) {
+        public void setMissedIngredientCount(Integer missedIngredientCount) {
             this.missedIngredientCount = missedIngredientCount;
         }
-
         public void setTitle(String title) {
             this.title = title;
         }
-
-        public void setUsedIngredientCount(float usedIngredientCount) {
+        public void setUsedIngredientCount(Integer usedIngredientCount) {
             this.usedIngredientCount = usedIngredientCount;
         }
 }
